requirements.txt
requests==2.25.1
python-dotenv==0.15.0
python-telegram-bot==13.1
google-cloud-dialogflow==2.0.0
dialogflow==1.1.0
vk-api==11.9.1